Made the NC pins on MPU-9250 passive.
NC is not really the right choice as we are allowed to route through
those pins. If we set the pins to NC they are a hard indication to ERC
that one must not connect anything to them. We need to make the decision
if we leave the pins unconnected on the schematic level instead.