Added a mechanical footprint for a 2.8" tft display with cap touch.
This footprint does not contain any electrical parts only mechanical
information and drawings including the registration holes for the
display, and where recommended cutouts should go to feed the ribbon
cables to the reverse side of the PCB. The drawing is based on the
buydisplay.com ER-TFT028-4 display with ER-TPC028-4.1 capacitive touch
sensor. I have also merged the outlines found on Adafruit breakouts to
hopefully allow the same footprint to be used with her displays. (sadly
Adafruit does not share information where she is getting her displays or
selling her displays without breakout boards, but if you already have a
display from her this footprint should work :TM: ) The connector library
contains the 50pin and 6pin connector footprints that can be placed
relative to this mechanical footprint. The positions where they should
go are marked in writing on this footprint.